requirements .txt
Flask==2.3.3
tflite-runtime==2.5.0
numpy==1.24.3
Pillow==9.5.0
